Created a greybox of the map
I have added the street, the buildings and some possible ideas such as a little playground and a big landmark, which would convey awe as it is the first thing that the player would see when getting out of the enclosed space. I added buildings in the background as well, in order to give the illusion that the world is bigger than it actually is. Needs further iterating, but it looks like a promising start