Updated the level and more.
I have updated the level and added the strip club and casino area. Some assets are still missing, and when they are ready they will be added to the level. I have also added the particles for enemy projectiles and made them shoot. They still need some work, but they should be fine for the showcase. Also moved the game project into the game folder.